+++ src/App.tsx
@@ -5,7 +5,7 @@ import { SidebarProvider } from "@context/sidebarContext";
 import { TodoContextProvider } from "@context/todoContext";
 import { AppDataContextProvider } from "@context/appDataContext";
 import Main from "@compo/mainApp";
-import { SelectedCategoryContextProvider } from "@context/selectCategoryContext";
+ 
 import CustomeRouter from "@services/customeRouter";
 import CustomeHistory from "@services/customeHistory";
 
@@ -16,16 +16,12 @@ function App() {
     <CustomeRouter history={CustomeHistory}>
       <ThemeContextProvider>
         <AppDataContextProvider>
-          <SelectedCategoryContextProvider>
-            
               <SidebarProvider>
                 <TodoContextProvider>
                   <Toaster position="bottom-center" reverseOrder={true} />
                   <Main />
                 </TodoContextProvider>
               </SidebarProvider>
-            
-          </SelectedCategoryContextProvider>
         </AppDataContextProvider>
       </ThemeContextProvider>
     </CustomeRouter>
