+++ src/components/sidebar/sidebar.tsx
@@ -4,15 +4,14 @@ import { FiChevronDown, FiChevronRight } from "react-icons/fi";
 import { RiRestaurant2Line } from "react-icons/ri";
 import { Link } from "react-router-dom";
 import { TodoContext } from "../../context/todoContext";
-import Axios   from "../../services/api";
-import {AppDataContext} from '@context/appDataContext' 
-import { SelectedCategoryContext } from "../../context/selectCategoryContext";
+import Axios from "../../services/api";
+import { AppDataContext } from "@context/appDataContext";
 
 // import "./sidebarStyles.css";
 import Toast from "../../util/toast";
 const Sidebar = () => {
   const theme = useContext(themeContext);
-  const { newCategorySelected ,selected } = useContext(SelectedCategoryContext);
+
   const {
     show,
     setThreeColAll,
@@ -23,20 +22,22 @@ const Sidebar = () => {
     setTableDone,
   } = useContext(TodoContext);
 
-  
-  const {updateCategory,
-    updateCategoryOff}  = useContext(AppDataContext)
-
+  const {
+    updateCategory,
+    updateCategoryOff,
+    newCategorySelected,
+    selected,
+    todoList,
+  } = useContext(AppDataContext);
 
   const sidebarStyle = {
     backgroundColor: theme.sidebar,
   };
- 
+
   const subsetTabsStyle = {
     color: theme.text3,
-   
   };
- 
+
   type ITodoState = "all" | "done";
   const [todoState, setTodoState] = useState<ITodoState>("all");
   const [categoryList, setCategoryList] = useState([]);
@@ -61,74 +62,49 @@ const Sidebar = () => {
     }
   }, [updateCategory]);
 
-  const showTodos = (id: string) => {
-    switch (id) {
-      case "all": {
-        setTodoState("all");
-        if (show[0] === "3col") setThreeColAll();
-        if (show[0] === "1col") setOneColAll();
-        if (show[0] === "table") setTableAll();
-        break;
-      }
-      case "done": {
-        setTodoState("done");
-
-        if (show[0] === "3col") setThreeColDone();
-        if (show[0] === "1col") setOneColDone();
-        if (show[0] === "table") setTableDone();
-        break;
-      }
-      case "3col": {
-        todoState === "all" ? setThreeColAll() : setThreeColDone();
-
-        break;
-      }
-      case "1col": {
-        todoState === "all" ? setOneColAll() : setOneColDone();
-
-        break;
-      }
-      case "table": {
-        todoState === "all" ? setTableAll() : setTableDone();
-        break;
-      }
-      default:
-        setThreeColAll();
-    }
-  };
-
   return (
     <div id="sidebar" style={sidebarStyle}>
-     
-          {categoryList.length ? (
-            <ul id="listCategories">
-              {categoryList.map((item) => (
-                <li
-                  key={item.uuid}
-                  className={`list-category-items ${selected === item.uuid ? "active-item" : ""}  `}
-                  style={subsetTabsStyle}
-                  onClick={() => {
-                    newCategorySelected(item.uuid , item.title);
-                  }}
-                >
-                  <div className="task-title-style">{item.title}</div>
-                  <div
-                    style={{
-                      width:
-                        item.task_count.length > 1 ? "fit-content" : "1.2rem",
-                    }}
-                    className="task-count-style"
-                  >
-                    {item.task_count}
-                  </div>
-                </li>
-              ))}
-            </ul>
-          ) : (
-            <span></span>
-          )}
-        </div>
+      {categoryList.length ? (
+        <ul id="listCategories">
+          <li
+            className={`list-category-items ${
+              selected === "other" ? "active-item" : ""
+            }  `}
+            style={subsetTabsStyle}
+            onClick={() => {
+              newCategorySelected();
+            }}
+          >
+            <div className="task-title-style">All</div>
 
+          </li>
+          {categoryList.map((item) => (
+            <li
+              key={item.uuid}
+              className={`list-category-items ${
+                selected === item.uuid ? "active-item" : ""
+              }  `}
+              style={subsetTabsStyle}
+              onClick={() => {
+                newCategorySelected(item.uuid);
+              }}
+            >
+              <div className="task-title-style">{item.title}</div>
+              <div
+                style={{
+                  width: item.task_count.length > 1 ? "fit-content" : "1.2rem",
+                }}
+                className="task-count-style"
+              >
+                {item.task_count}
+              </div>
+            </li>
+          ))}
+        </ul>
+      ) : (
+        <span></span>
+      )}
+    </div>
   );
 };
