+++ src/styles/_drawer.scss
@@ -5,7 +5,9 @@
   background-color: var(--background);
   color: var(--text1);
   width: 100%;
-  height: 100vh;
+  max-height: 100vh;
+  min-height: 100vh;
+
   padding: 10px;
   display: flex;
   flex-direction: column;
@@ -24,20 +26,48 @@
 
     #text-area-footer {
       padding: 0.5rem 0.5rem;
+      
+      
+      
+      #button-edit-todo{
+        text-transform: capitalize;
+        letter-spacing: 1px;
+        font-weight: bold;
+        border-radius:10px;
+        border:1px solid var(--secondSidebar);
+        color:var(--text1);
+        &:hover{
+          background-color: var(--text1);
+          color:var(--secondSidebar)
+        }
+        
+        &:disabled {
+          // background-color: red;  
+          color:#4a4948;
+          border:1px solid #4a4948;
+        }
+
+        &:active{
+          border:5px solid var(--secondSidebar);
+        }
+
+
+      }
     }
   }
 
   #drawer-box-footer {
     width: 100%;
     padding: 10px;
-    // position:fixed;
-    // bottom:10px;
     display: flex;
     flex-wrap: no-wrap;
     gap: 1rem;
 
-    // background-color: gray;
-    // overflow-y:auto;
+
+
+
+
+ 
     .drawer-icon-box {
       display: flex;
       justify-content: center;
