+++ src/styles/_modals.scss
@@ -131,15 +131,15 @@
 
 .Modal_Radio_Container_Dark > label {
   color: white;
-margin-right: auto;
-padding: .4rem 0;
-font-size:1rem;
+  margin-right: auto;
+  padding: .4rem 0;
+  font-size:1rem;
 }
 .Modal_Radio_Container_Light > label {
   color: black;
-margin-right: auto;
-padding: .4rem 0;
-font-size:1rem;
+  margin-right: auto;
+  padding: .4rem 0;
+  font-size:1rem;
 }
