+++ src/util/toast.tsx
@@ -3,7 +3,7 @@ import {useContext} from "react"
 import { toast } from "react-hot-toast";
 // const ThemeContext from "../context/colorModeContext"
 
-export default function Toast (msg,success=true){
+export default function Toast (msg : string,success : boolean =true){
